bms-monitor/requirements.txt
# FastAPI 核心依賴
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# WebSocket 支援
websockets>=11.0

# 數據庫與 ORM
sqlalchemy>=2.0.0
alembic>=1.12.0
aiosqlite>=0.19.0

# 緩存
redis>=5.0.0
aioredis>=2.0.0

# MQTT 客戶端
paho-mqtt>=1.6.0

# 數據處理
pydantic>=2.4.0
pydantic-settings>=2.0.0

# 工具庫
python-dotenv>=1.0.0
structlog>=23.1.0
asyncio-mqtt>=0.13.0

# 藍牙通訊 (現有 BMS 模組)
bleak>=0.21.0

# HTTP 客戶端
httpx>=0.25.0

# 日期時間處理
python-dateutil>=2.8.0

# JSON 序列化
orjson>=3.9.0